Exclude FieldNames from public enum.
Change index field name definition to package access.
