+++ airsonic-main/src/main/java/org/airsonic/player/service/search/IndexType.java
@@ -75,78 +75,6 @@ public enum IndexType {
 
     ;
 
-  public static final class FieldNames {
-
-    private FieldNames() {
-    }
-
-    ;
-
-    /**
-     * A field same to a legacy server, id field.
-     * 
-     * @since legacy
-     **/
-    public static final String ID = "id";
-
-    /**
-     * A field same to a legacy server, id field.
-     * 
-     * @since legacy
-     **/
-    public static final String FOLDER_ID = "fId";
-
-    /**
-     * A field same to a legacy server, numeric field.
-     * 
-     * @since legacy
-     **/
-    public static final String YEAR = "y";
-
-    /**
-     * A field same to a legacy server, key field.
-     * 
-     * @since legacy
-     **/
-    public static final String GENRE = "g";
-
-    /**
-     * A field same to a legacy server, key field.
-     * 
-     * @since legacy
-     **/
-    public static final String MEDIA_TYPE = "m";
-
-    /**
-     * A field same to a legacy server, key field.
-     * 
-     * @since legacy
-     **/
-    public static final String FOLDER = "f";
-
-    /**
-     * A field same to a legacy server, usually with common word parsing.
-     * 
-     * @since legacy
-     **/
-    public static final String ARTIST = "art";
-
-    /**
-     * A field same to a legacy server, usually with common word parsing.
-     * 
-     * @since legacy
-     **/
-    public static final String ALBUM = "alb";
-
-    /**
-     * A field same to a legacy server, usually with common word parsing.
-     * 
-     * @since legacy
-     **/
-    public static final String TITLE = "tit";
-
-  }
-
   @SafeVarargs
   private static final Map<String, Float> boosts(SimpleEntry<String, Float>... entry) {
     Map<String, Float> m = new HashMap<>();
