+++ airsonic-main/src/test/java/org/airsonic/player/service/search/AnalyzerFactoryTestCase.java
@@ -9,7 +9,6 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import org.airsonic.player.service.search.IndexType.FieldNames;
 import org.apache.lucene.analysis.TokenStream;
 import org.apache.lucene.analysis.tokenattributes.TermAttribute;
 import org.junit.Test;
